Extract YouTube transcript into plugins/youtube/
- Move lib/youtube.js → plugins/youtube/youtube.js
- Create plugins/youtube/index.js with register(app, ctx)
- Create plugins/youtube/apt.txt (python3-minimal)
- Remove yt-dlp + python3 from base Dockerfile, add to with-plugins stage
- Wire plugin system into server.js: imports, pluginEvents, auth middleware,
  expanded pluginCtx (withUserLimit, safePageClose, normalizeUserId,
  validateUrl, safeError, buildProxyUrl, proxyPool, failuresTotal),
  28 event emissions across all route handlers
- Update test import path
