+++ plugins/youtube/apt.txt
@@ -1,0 +1,1 @@
+python3-minimal
